#105 / refactor(sidebar) : 앱 셸과 사이드바 조합 책임을 app 레이어로 이동
