+++ src/app/(app)/layout.tsx
@@ -1,4 +1,4 @@
-import { AppShell } from "@/shared/layout/AppShell";
+import { AppShell } from "@/app/(app)/_components/AppShell";
 
 export default function AppLayout({ children }: { children: React.ReactNode }) {
   return <AppShell>{children}</AppShell>;
